Add dev:e2e command, which spins up the test environment without running tests
and delegate to a common script across dev and test commands
