+++ e2e-tests/package.json
@@ -5,11 +5,15 @@
   "main": "index.js",
   "type": "module",
   "scripts": {
+    "dev:e2e": "node --experimental-strip-types --disable-warning=ExperimentalWarning dev.ts",
     "test": "bddgen && playwright test",
     "test:headed": "bddgen && playwright test --headed",
     "test:report": "playwright show-report",
     "test:ui": "bddgen && playwright test --ui-port 6080",
     "test:log": "bddgen && DEBUG=testcontainers* playwright test --ui-port 6080"
+  },
+  "engines": {
+    "node": ">=22.12"
   },
   "keywords": [],
   "author": "",
